+++ activerecord/test/cases/relation/select_test.rb
@@ -11,5 +11,10 @@ module ActiveRecord
       expected = Post.select(:title).to_sql
       assert_equal expected, Post.select(nil).select(:title).to_sql
     end
+
+    def test_reselect
+      expected = Post.select(:title).to_sql
+      assert_equal expected, Post.select(:title, :body).reselect(:title).to_sql
+    end
   end
 end